@@ -7,18 +7,21 @@
     "node_modules"
   ],
   "dependencies": {
-    "@angular/common": "5.2.8",
-    "@angular/compiler": "5.2.8",
-    "@angular/core": "5.2.8",
-    "@angular/forms": "5.2.8",
-    "@angular/platform-browser": "5.2.8",
-    "@angular/platform-browser-dynamic": "5.2.8",
-    "@angular/router": "5.2.8",
-    "@ng-bootstrap/ng-bootstrap": "1.0.1",
-    "bootstrap": "4.0.0",
+    "@angular/cdk": "^5.2.5",
+    "@angular/common": "6.0.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/core": "6.0.0",
+    "@angular/forms": "6.0.0",
+    "@angular/material": "^5.2.5",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/router": "6.0.0",
+    "@ng-bootstrap/ng-bootstrap": "1.1.1",
+    "bootstrap": "4.1.2",
     "core-js": "2.5.3",
     "font-awesome": "4.7.0",
     "jquery": "3.3.1",
+    "ng-jhipster": "^0.6.0",
     "ngx-cookie": "2.0.1",
     "ngx-infinite-scroll": "0.5.1",
     "ngx-webstorage": "2.0.1",
@@ -30,8 +33,8 @@
     "zone.js": "0.8.20"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.3",
-    "@angular/compiler-cli": "5.2.8",
+    "@angular/cli": "^7.1.4",
+    "@angular/compiler-cli": "6.0.0",
     "@ngtools/webpack": "6.0.0-beta.4",
     "@types/jasmine": "2.8.6",
     "@types/node": "9.4.7",
@@ -49,12 +52,12 @@
     "extract-text-webpack-plugin": "4.0.0-beta.0",
     "file-loader": "1.1.11",
     "fork-ts-checker-webpack-plugin": "0.4.2",
-    "generator-NgMcmonitor": "4.14.0",
+    "generator-jhipster": "^5.7.2",
     "html-loader": "0.5.5",
     "html-webpack-plugin": "3.0.6",
     "husky": "0.14.3",
     "jasmine-core": "2.7.0",
-    "karma": "1.7.1",
+    "karma": "^3.1.4",
     "karma-chrome-launcher": "2.2.0",
     "karma-coverage": "^1.1.2",
     "karma-intl-shim": "1.0.3",
