@@ -3,11 +3,11 @@ import { FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 import { CommonModule } from '@angular/common';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { NgMcmonitorModule } from 'ng-mcmonitor';
+import { McmonitorModule } from 'app/shared/mcmonitor';
 import { InfiniteScrollModule } from 'ngx-infinite-scroll';
 
 @NgModule({
-    imports: [NgbModule.forRoot(), NgMcmonitorModule.forRoot({}), InfiniteScrollModule],
-    exports: [FormsModule, HttpClientModule, CommonModule, NgbModule, NgMcmonitorModule, InfiniteScrollModule]
+    imports: [NgbModule.forRoot(), McmonitorModule.forRoot(), InfiniteScrollModule],
+    exports: [FormsModule, HttpClientModule, CommonModule, NgbModule, McmonitorModule, InfiniteScrollModule]
 })
 export class MicroMonitorRegistrySharedLibsModule {}
