@@ -1,4 +1,4 @@
-import { McmEventManager } from 'ng-mcmonitor';
+import { McmEventManager } from 'app/shared/mcmonitor';
 import { HttpInterceptor, HttpRequest, HttpErrorResponse, HttpHandler, HttpEvent } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/do';
